requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.12.14
click==8.1.7
colorama==0.4.6
distro==1.9.0
eval_type_backport==0.2.0
fastapi==0.115.5
griffe==1.5.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
logfire-api==2.7.1
mangum==0.19.0
openai==1.57.4
pydantic==2.10.1
pydantic-ai-slim==0.0.13
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.32.1
